new test of esr

+++ src/tracking/rfcavity.jl
@@ -1,7 +1,7 @@
 include("drift.jl")
 
 function RFCavityPass!(r_in::Array{Float64,1}, le::Float64, nv::Float64, freq::Float64, h::Float64, 
-    lag::Float64, philag::Float64, nturn::Int, T0::Int, num_particles::Int, lost_flags::Array{Int64,1})
+    lag::Float64, philag::Float64, nturn::Int, T0::Float64, num_particles::Int, lost_flags::Array{Int64,1})
     # le - physical length
     # nv - peak voltage (V) normalized to the design enegy (eV)
     # freq - frequency (Hz)
